Fix various compilation errors in snippets
