+++ apps_script/execute/execute.go
@@ -16,11 +16,92 @@
  */
 package main
 
+import (
+	"encoding/json"
+	"fmt"
+	"golang.org/x/oauth2/google"
+	"io/ioutil"
+	"log"
+	"net/http"
+	"os"
+	"golang.org/x/net/context"
+	"golang.org/x/oauth2"
+	"google.golang.org/api/script/v1"
+)
+
+// Retrieve a token, saves the token, then returns the generated client.
+func getClient(config *oauth2.Config) *http.Client {
+	// The file token.json stores the user's access and refresh tokens, and is
+	// created automatically when the authorization flow completes for the first
+	// time.
+	tokFile := "token.json"
+	tok, err := tokenFromFile(tokFile)
+	if err != nil {
+		tok = getTokenFromWeb(config)
+		saveToken(tokFile, tok)
+	}
+	return config.Client(context.Background(), tok)
+}
+
+// Request a token from the web, then returns the retrieved token.
+func getTokenFromWeb(config *oauth2.Config) *oauth2.Token {
+	authURL := config.AuthCodeURL("state-token", oauth2.AccessTypeOffline)
+	fmt.Printf("Go to the following link in your browser then type the "+
+		"authorization code: \n%v\n", authURL)
+
+	var authCode string
+	if _, err := fmt.Scan(&authCode); err != nil {
+		log.Fatalf("Unable to read authorization code: %v", err)
+	}
+
+	tok, err := config.Exchange(context.TODO(), authCode)
+	if err != nil {
+		log.Fatalf("Unable to retrieve token from web: %v", err)
+	}
+	return tok
+}
+
+// Retrieves a token from a local file.
+func tokenFromFile(file string) (*oauth2.Token, error) {
+	f, err := os.Open(file)
+	if err != nil {
+		return nil, err
+	}
+	defer f.Close()
+	tok := &oauth2.Token{}
+	err = json.NewDecoder(f).Decode(tok)
+	return tok, err
+}
+
+// Saves a token to a file path.
+func saveToken(path string, token *oauth2.Token) {
+	fmt.Printf("Saving credential file to: %s\n", path)
+	f, err := os.OpenFile(path, os.O_RDWR|os.O_CREATE|os.O_TRUNC, 0600)
+	if err != nil {
+		log.Fatalf("Unable to cache oauth token: %v", err)
+	}
+	defer f.Close()
+	json.NewEncoder(f).Encode(token)
+}
+
+
 func main() {
 	// [START apps_script_api_execute]
 
 	scriptId := "ENTER_YOUR_SCRIPT_ID_HERE"
-	client := getClient(ctx, config)
+
+	b, err := ioutil.ReadFile("credentials.json")
+	if err != nil {
+		log.Fatalf("Unable to read client secret file: %v", err)
+	}
+
+	// If modifying these scopes, delete your previously saved token.json.
+	config, err := google.ConfigFromJSON(b, "https://www.googleapis.com/auth/script.projects")
+	if err != nil {
+		log.Fatalf("Unable to parse client secret file to config: %v", err)
+	}
+
+	client := getClient(config)
 
 	// Generate a service object.
 	srv, err := script.New(client)
@@ -39,19 +120,20 @@ func main() {
 	}
 
 	if resp.Error != nil {
-		// The API executed, but the script returned an error.
+		// The API executed, but the script returned an details.
 
-		// Extract the first (and only) set of error details and cast as a map.
+		// Extract the first (and only) set of details details and cast as a map.
 		// The values of this map are the script's 'errorMessage' and
 		// 'errorType', and an array of stack trace elements (which also need to
 		// be cast as maps).
-		error := resp.Error.Details[0].(map[string]interface{})
-		fmt.Printf("Script error message: %s\n", error["errorMessage"])
+		var details map[string]interface{};
+		json.Unmarshal(resp.Error.Details[0], &details)
+		fmt.Printf("Script details message: %s\n", details["errorMessage"])
 
-		if error["scriptStackTraceElements"] != nil {
+		if details["scriptStackTraceElements"] != nil {
 			// There may not be a stacktrace if the script didn't start executing.
-			fmt.Printf("Script error stacktrace:\n")
-			for _, trace := range error["scriptStackTraceElements"].([]interface{}) {
+			fmt.Printf("Script details stacktrace:\n")
+			for _, trace := range details["scriptStackTraceElements"].([]interface{}) {
 				t := trace.(map[string]interface{})
 				fmt.Printf("\t%s: %d\n", t["function"], int(t["lineNumber"].(float64)))
 			}
@@ -61,7 +143,8 @@ func main() {
 		// based upon what types the Apps Script function returns. Here, the
 		// function returns an Apps Script Object with String keys and values, so
 		// must be cast into a map (folderSet).
-		r := resp.Response.(map[string]interface{})
+		var r map[string]interface{};
+		json.Unmarshal(resp.Response, &r)
 		folderSet := r["result"].(map[string]interface{})
 		if len(folderSet) == 0 {
 			fmt.Printf("No folders returned!\n")
