+++ classroom/snippets/getCourse.go
@@ -17,69 +17,27 @@
 package main
 
 import (
-        "encoding/json"
         "fmt"
+        "golang.org/x/oauth2/google"
+        "google.golang.org/api/classroom/v1"
         "io/ioutil"
         "log"
         "net/http"
-        "os"
-
-        "golang.org/x/net/context"
-        "golang.org/x/oauth2"
-        "golang.org/x/oauth2/google"
-        "google.golang.org/api/classroom/v1"
 )
 
-// Retrieve a token, saves the token, then returns the generated client.
-func getClient(config *oauth2.Config) *http.Client {
-        tokFile := "token.json"
-        tok, err := tokenFromFile(tokFile)
+func getCourse(client *http.Client) {
+        // [START classroom_get_course]
+        srv, err := classroom.New(client)
         if err != nil {
-                tok = getTokenFromWeb(config)
-                saveToken(tokFile, tok)
+                log.Fatalf("Unable to create classroom Client %v", err)
         }
-        return config.Client(context.Background(), tok)
-}
-
-// Request a token from the web, then returns the retrieved token.
-func getTokenFromWeb(config *oauth2.Config) *oauth2.Token {
-        authURL := config.AuthCodeURL("state-token", oauth2.AccessTypeOffline)
-        fmt.Printf("Go to the following link in your browser then type the "+
-                "authorization code: \n%v\n", authURL)
-
-        var authCode string
-        if _, err := fmt.Scan(&authCode); err != nil {
-                log.Fatalf("Unable to read authorization code: %v", err)
-        }
-
-        tok, err := config.Exchange(context.TODO(), authCode)
+        id := "123456"
+        course, err := srv.Courses.Get(id).Do()
         if err != nil {
-                log.Fatalf("Unable to retrieve token from web: %v", err)
+                log.Fatalf("Course unable to be retrieved %v", err)
         }
-        return tok
-}
-
-// Retrieves a token from a local file.
-func tokenFromFile(file string) (*oauth2.Token, error) {
-        f, err := os.Open(file)
-        if err != nil {
-                return nil, err
-        }
-        defer f.Close()
-        tok := &oauth2.Token{}
-        err = json.NewDecoder(f).Decode(tok)
-        return tok, err
-}
-
-// Saves a token to a file path.
-func saveToken(path string, token *oauth2.Token) {
-        fmt.Printf("Saving credential file to: %s\n", path)
-        f, err := os.OpenFile(path, os.O_RDWR|os.O_CREATE|os.O_TRUNC, 0600)
-        if err != nil {
-                log.Fatalf("Unable to cache oauth token: %v", err)
-        }
-        defer f.Close()
-        json.NewEncoder(f).Encode(token)
+        // [END classroom_get_course]
+        fmt.Printf("Course with ID %v found.", course.Id)
 }
 
 func main() {
@@ -94,16 +52,5 @@ func main() {
                 log.Fatalf("Unable to parse client secret file to config: %v", err)
         }
         client := getClient(config)
-        // [START classroom_get_course]
-        srv, err := classroom.New(client)
-        if err != nil {
-                log.Fatalf("Unable to create classroom Client %v", err)
-        }
-        id := "123456"
-        course, err := srv.Courses.Get(id).Do()
-        if err != nil {
-                log.Fatalf("Course unable to be retrieved %v", err)
-        }
-        // [END classroom_get_course]
-        fmt.Printf("Course with ID %v found.", course.Id)
+        getCourse(client)
 }
