+++ slides/snippets/oauth.go
@@ -1,6 +1,7 @@
 package snippets
 
 import (
+	"log"
 	"golang.org/x/net/context"
 	"golang.org/x/oauth2/google"
 	"google.golang.org/api/drive/v2"
